Use extended loggers intead of proxy
