+++ logback-ext/src/main/java/com/tersesystems/logback/ext/LoggerStatement.java
@@ -8,7 +8,7 @@
  *
  *     http://creativecommons.org/publicdomain/zero/1.0/
  */
-package com.tersesystems.logback.proxy;
+package com.tersesystems.logback.ext;
 
 import org.slf4j.Marker;
 import org.slf4j.Logger;
