+++ settings.gradle
@@ -27,4 +27,4 @@ includeProject 'logback-core'
 includeProject 'logback-example'
 includeProject 'logback-guice-example'
 includeProject 'logstash-logback-context'
-includeProject 'logback-proxy'
+includeProject 'logback-ext'
